pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fida"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Core dependencies with compatible versions
    "loguru>=0.7.0",
    "lxml>=4.9.0",
    "matplotlib>=3.8.0",
    "nest-asyncio>=1.5.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pandas-datareader>=0.10.0",
    "pyrate-limiter==2.6.0",
    "requests-cache>=1.1.0",
    "requests-ratelimiter>=0.4.0,<0.8.0",
    "scipy>=1.12.0",
    "setuptools>=69.0.0",
    "tqdm>=4.66.0",
    "tiingo>=0.14.0",
    "yfinance>=0.2.0"
]

[project.optional-dependencies]
test = [
    "coveralls>=3.3.1",
    "freezegun>=1.2.0",
    "requests-mock>=1.11.0",
    "pytest>=7.0.0"
]

dev = [
    "better-exceptions>=0.3.3",
    "ruff>=0.2.0",
    "ipython>=8.12.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "pre-commit>=3.5.0",
    "terminaltables>=3.1.10",
    "tabulate>=0.9.0",
    "types-ujson>=5.9.0",
    # Type stubs
    "pandas-stubs>=2.2.0",
    "types-requests>=2.31.0.20240125",
    "types-setuptools>=69.0.0",
    "types-tqdm>=4.66.0",
    "types-urllib3>=1.26.25.14"
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "RUF"  # ruff-specific rules
]

exclude = [
    ".git",
    ".ruff_cache",
    "build",
    "dist"
]

[tool.ruff.lint.isort]
combine-as-imports = true
